requirements.txt
-r requirements-dev.txt
gunicorn
psycopg2

Django~=3.2.7
django-widget-tweaks~=1.4.8
django-allauth~=0.45.0
pandas~=1.3.0
scikit-learn~=0.24.2
Pillow~=8.3.1
xhtml2pdf~=0.2.5
reportlab~=3.6.1